# `===` Optimization: Defer recursion into pointees