@JeffBezanson 2022-01-04
Good idea, I like it! This sort of feels like the right way to do it to me, even if it's a tad slower in some cases.

I wonder if it would help to ignore pointer fields completely in the first pass, instead of doing the `if (af != bf && (af == NULL || bf == NULL))` check?