@nanosoldier 2022-01-05
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/43658#issuecomment-1005111658):
```
NanosoldierError: error when preparing/pushing to report repo: failed process: Process(setenv(`git push`; dir="/data/nanosoldier/workdir/NanosoldierReports"), ProcessExited(1)) [1]

```
Unfortunately, the logs could not be uploaded.