@NHDaly 2022-01-05
Yay, thanks! 🎉 

> I wonder if it would help to ignore pointer fields completely in the first pass, instead of doing the `if (af != bf && (af == NULL || bf == NULL))` check?

Yeah, i had that idea too, and I could go either way. I chose to do it this way, because it's possible the NULL/NULL check could also allow us to return early: If a struct has 2 pointer fields and the _second_ one is NULL in `a` but not `b`, you could avoid the recursive call to compare the the first pointee if you had seen the NULL pointer first.

So it's another instance of front-loading checking the bits in the struct before recursing to pointees. So i think probably better to do it this way, since it's making the same tradeoff as the rest of the PR. Does that make sense to you, too?