@KristofferC 2022-01-05
> Anyone know why Nanosolider failed to upload the benchmark results? Is it because the PR is a draft or something?

The size has just gotten too big and no one has fixed it.