@JeffBezanson 2022-01-05
NULL fields are rare, so I think not worth special-casing. Actual timing is what matters though.