@JeffBezanson 2022-01-05
I get these results for your last benchmark (where the PR is slower):

before PR:
  132.834 μs (0 allocations: 0 bytes)
  124.298 μs (0 allocations: 0 bytes)

PR:
  192.262 μs (0 allocations: 0 bytes)
  193.349 μs (0 allocations: 0 bytes)

PR + ignore pointer fields on the first pass:
  180.625 μs (0 allocations: 0 bytes)
  183.335 μs (0 allocations: 0 bytes)

So there is a slight improvement (and looks like I need a new laptop again :joy: )