@NHDaly 2022-01-10
Thanks @vtjnash!

What do the percentages mean?

It looks like some got quite a bit better and some quite a bit worse, but in ways that don't exactly make sense to me... So I feel like probably this is mostly just noisy? Can any of you with more experience reading these weigh in?
<img width="1053" alt="Screen Shot 2022-01-10 at 12 08 53 PM" src="https://user-images.githubusercontent.com/1582097/148808131-2f5e2864-b901-4aa0-8491-58789786f48c.png">


I'll give your suggestion a shot now, @JeffBezanson.