@oscardssmith 2022-01-10
Nanosoldier (especially recently) is pretty noisy.