@KristofferC 2022-01-10
@NHDaly, you need to code quote the part after the nanosoldier invocation.