@oscardssmith 2022-01-11
it did work, it's just that there's a nanosoldier bug that means @vtjnash needs to post the log manually.