@vtjnash 2022-01-24
For more impact, we may want to update codegen.cpp to do the same ordering change as here also for `emit_bits_compare`