@nanosoldier 2022-01-25
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/43658#issuecomment-1020659872):
```
NanosoldierError: error when preparing/pushing to report repo: failed process: Process(setenv(`git push`; dir="/nanosoldier/workdir/NanosoldierReports"), ProcessExited(1)) [1]

```
Unfortunately, the logs could not be uploaded.