@vtjnash 2022-01-26
https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/3eac22a_vs_master/report.md

Still not what I mean to run though: @nanosoldier `runbenchmarks(!"scalar", vs=":master")`