@nanosoldier 2022-01-26
[Your benchmark job](https://github.com/JuliaLang/julia/pull/43658#issuecomment-1022369669) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/dd9418f_vs_051ab3b/report.md).